move event/cmd collectors to collectors dir
